Add files via upload
This Python code implements a function `pick_a_card()` that randomly selects a playing card from a standard deck of 52 cards. The deck consists of four suits (Diamonds, Spades, Hearts, Clubs) and ranks ranging from 2 to 10, along with face cards (Jack, Queen, King) and Ace.

The function `pick_a_card()` randomly chooses a card suit and rank and returns a string describing the selected card in the format "The [Rank] of [Suit]". Finally, the code demonstrates the usage of the function by printing out a randomly picked card.

Overall, this code simulates the action of picking a random card from a standard deck of playing cards.